Create Project2.py
In this project, I carried out mathematical and programming solutions to the probabilistic
calculation problems raised in the project. First, for the first part, I calculated the occurrence
times of Q, S and C, and compared them with the total number to get their respective
probabilities. In the second part, I calculated the number of times each score appeared, compared
it with the total number, and got their respective probability. In the third part, I used the results
obtained in the second part and presented the results with Venn. In the program, there is also a
third part of the code withdrawal. In the fourth part, I defined each time separately, expressed it
with mathematical formulas, and calculated the result with codes.